Switch from Original) Constructing parent language terms from recursive sublanguage terms To Different) Constructing parent language commands from recursive subanguage commands, splitting all sublanguages with regular command-handler pairings into recursive sublanguages, and using constuctors to wrap the commands of sublanguages as commands of the parent language.
Show that one's still prevented from deriving the extension from
recursive sublanguage programs to parent language programs.

This time it's because `hoistFree` requires an instance of functor
for the parent language's commands, and that this can't be
implemented in terms of the functor instances for the recursive
sublanguages due to a failure to unify the parent language term
variable in the (fmap)'s signature with examples in patterns.

This would be solved by allowing a dependent (fmap) to be used
to recurse into the (fmap) of sublanguages'. Hence, the new approach
seems to have an obscured need for dependent functions.

Is it viable to alternate with less-automated implementations?:
1) co/presheaves over languages giving automatically paired command and
handler types, encompassing the maximal nonrecursing sublanguages
2) datatypes of recursive sublanguages conjoined into coproducts
separately from products and paired by manually matching
coproduct (command) sections to product (handler) projections